+++ src/main/java/com/google/api/generator/engine/ast/AssignmentExpr.java
@@ -39,7 +39,7 @@ public abstract class AssignmentExpr implements Expr {
       TypeNode lhsType = assignmentExpr.variableExpr().variable().type();
       TypeNode rhsType = assignmentExpr.valueExpr().type();
       if (lhsType.isPrimitiveType()) {
-        if (rhsType == null) {
+        if (rhsType == TypeNode.NULL) {
           throw new TypeMismatchException(
               String.format(
                   "Null cannot be assigned to the primitive type %s", lhsType.toString()));
@@ -50,7 +50,8 @@ public abstract class AssignmentExpr implements Expr {
                   "LHS type %s must match RHS type %s", lhsType.toString(), rhsType.toString()));
         }
       } else {
-        if (!lhsType.reference().clazz().isAssignableFrom(rhsType.reference().clazz())) {
+        if (rhsType != TypeNode.NULL
+            && !lhsType.reference().clazz().isAssignableFrom(rhsType.reference().clazz())) {
           throw new TypeMismatchException(
               String.format(
                   "LHS type %s must be a supertype of the RHS type %s",
