+++ src/main/java/com/google/api/generator/engine/ast/TypeNode.java
@@ -37,6 +37,8 @@ public abstract class TypeNode implements AstNode {
   public static final TypeNode VOID = builder().setTypeKind(TypeKind.VOID).build();
   public static final TypeNode INT = builder().setTypeKind(TypeKind.INT).build();
   public static final TypeNode BOOLEAN = builder().setTypeKind(TypeKind.BOOLEAN).build();
+  public static final TypeNode NULL =
+      withReference(Reference.withClazz(javax.lang.model.type.NullType.class));
   public static final TypeNode STRING = withReference(Reference.withClazz(String.class));
   public static final TypeNode STRING_ARRAY =
       builder()
